Fixing race condition while handling lots of data
The callback batman_data_ready() set the variable data_ready_cond to
true before waking up the packet handling thread. When the thread has
completed its work it will sleep and sets the variable back to false.
When the system has to handle large quantities of traffic the callback
might be called several times while the packet handling thread runs. The
status variable will be overwritten once the thread is done which might
leave packets in the socket buffer. If the socket buffer is full no more
packets are accepted and the callback won't be triggered. Therefore the
packet receive thread has to reset the variable at the very beginning.