Fixes rounding issues in vis.c
This patches fixes two rounding issues in vis.c for sending and purging
vis packets. Before, the timers and timeouts always got rounded down to
seconds, though we want a precision in milliseconds.

This also fixes a kernel panic that occures when lowering the timer for
sending vis packets (vis_interval) to less than 1000ms in main.c
manually.